move material export bridge func
