+++ Packages/net.yutopp.vgltf.ext.vrm0.unity/Runtime/DefaultMaterialExporterBrigde.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fef5335f0cc0d4fd2831cf6c4b61a078
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
